+++ CHANGES.md
@@ -19,7 +19,10 @@ Beta Releases
                 destination : Cesium.Cartesian3.fromDegrees(-117.16, 32.71, 15000.0)
             });
 
-   * Renamed `options.endReferenceFrame` to `options.endTransform` in `Camera.flyTo` and `Camera.flyToRectangle`.
+   * In `Camera.flyTo` and `Camera.flyToRectangle`, renamed:
+      * `options.endReferenceFrame` to `options.endTransform`.
+      * `options.onComplete` to `options.complete`.
+      * `options.onCancel` to `options.cancel`.
    * Renamed `Simon1994PlanetaryPositions` functions `ComputeSunPositionInEarthInertialFrame` and `ComputeMoonPositionInEarthInertialFrame` to `computeSunPositionInEarthInertialFrame` and `computeMoonPositionInEarthInertialFrame`, respectively.
    * Replaced `Scene.scene2D.projection` property with read-only `Scene.mapProjection`.  Set this with the `mapProjection` option for the `Viewer`, `CesiumWidget`, or `Scene` constructors.
    * `Scene` constructor function now takes an `options` parameter instead of individual parameters.
@@ -27,6 +30,8 @@ Beta Releases
    * CZML property references now use a `#` symbol to separate identifier from property path. `objectId.position` should now be `objectId#position`. 
    * `CesiumWidget.showErrorPanel` now takes a `message` parameter in between the previous `title` and `error` parameters.
    * `Event.removeEventListener` no longer throws `DeveloperError` if the `listener` does not exist; it now returns `false`.
+   * `AnimationCollection.add` renamed `options.onUpdate` to `options.update`.
+   * `AnimationCollection.add`, `AnimationCollection.addAlpha`, and `AnimationCollection.addProperty` renamed `options.onComplete` to `options.complete` and `options.onCancel` to `options.cancel`.
 * `DynamicObject.id` can now include period characters.
 * `ReferenceProperty` can now handle sub-properties, for example, `myObject#billboard.scale`.
 * Added `Cesium.VERSION` to the combined `Cesium.js` file.
@@ -40,7 +45,7 @@ Beta Releases
 * `Viewer` and `CesiumWidget` now take a new optional parameter, `creditContainer`.
 * Added `PerformanceWatchdog` widget and `viewerPerformanceWatchdogMixin`.
 * Fixed a problem that could rarely lead to the camera's `tilt` property being `NaN`.
-* Updated third-party [Tween.js](https://github.com/sole/tween.js/) from r7 to r13.
+* Updated third-party [Tween.js](https://github.com/sole/tween.js/) from r12 to r13.
 
 ### b29 - 2014-06-02